bitcoin core rpc tests included. python integration test completed
