+++ Cargo.toml
@@ -28,6 +28,7 @@ serde_json = "1.0.103"
 uniffi = { version = "0.24.3" }
 url = "2.4.0"
 thiserror = "1.0.47"
+bitcoincore-rpc = "0.17.0"
 
 [[bin]]
 name = "uniffi-bindgen"
